requirements.txt
Django>=4.2,<4.3
django-cors-headers==4.2.0
django-environ==0.10.0
django-proxy==1.2.2
geomanager==0.0.5
isodate==0.6.1
psycopg2==2.9.7
OWSLib==0.29.2
wagtail>=5.1,<5.2
wagtail-adminsortable @ https://github.com/Lh4cKg/wagtail-admin-sortable/archive/baa0d1f13a6bca4cc60e0d7dcd0522ecafcf5d67.zip#sha256=ab6ea62b180659f56330397e97e9fcc2769dd10801f52fd259bc3175d77e27b7